requirements.txt
django-kdl-wagtail
django-adminactions
django-compressor
django-ckeditor
django-extensions
django-geonames-place
django-kdl-ldap
django-languages-plus
django-libsass
django-model-utils
django-nested-admin
django-polymorphic
django-redis
django-rest-polymorphic
django-reversion
django-script-codes
pandas
psycopg2-binary
redis
xlrd